requirements.txt
Django==5.0.2
pydicom==2.4.3
Pillow==10.2.0
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10 